Update Grid.py
Added in the otherHalfCoordinates, otherHalfInfo and a new dictionary for detemringing the values at every orientation.  Most of playCard should function now, just needs to make sure that it works properly as there is one section on the use of updating the dictionarrie's value which is untested and needs to be checked.