Ai as good as it can be
The heuristic on this branch has been optimised as much as possible. The only way I can see to make it better is to increase the depth of the tree or figure out a different heuristic.
Increasing the depth makes regular play take >3 seconds though
Reduced recycling search depth to 1. Couldn't think of a way to reduce computation without making ai miss possibilities